backend/requirements.txt
# MTC Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
email-validator>=2.1.0

# Config
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# LLM & AI
openai>=1.12.0
tenacity>=8.2.3
numpy>=1.26.0

# Celery & Task Queue
celery>=5.3.0
redis>=5.0.0

# Storage
boto3>=1.34.0
minio>=7.2.0

# Utilities
python-dateutil>=2.8.0
emoji>=2.8.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Optional: KouriChat 微信支持
# wxautox4>=4.0.0
